fix: remove all rename @typescript-no-watch to no-watch
