+++ rules/no-watch/examples/correct.ts
