+++ rules/no-watch/examples/incorrect-with-effect.ts
