+++ rules/no-watch/examples/incorrect-with-event.ts
