+++ rules/no-watch/examples/incorrect-with-guard.ts
