+++ rules/no-watch/examples/incorrect-with-sample.ts
