+++ rules/no-watch/examples/incorrect-with-store.ts
